+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Fletch Changelog
 
+## 0.3.0
+_Not Yet Released_
+
+ * Added `.classes.contains()`, `.classes.containsAll()`,
+   `.classes.containsAny()`, `.classes.addAll()`, `.classes.removeAll()`, and
+   `.classes.toggleAll()`
+
 ## 0.2.0
 _Released 7/20/2014_
 
@@ -7,7 +14,8 @@ _Released 7/20/2014_
  
  * Added `.value` for getting/setting form input values
  
- * Added `.checked` for getting/setting checked state for checkboxes and radio buttons
+ * Added `.checked` for getting/setting checked state for checkboxes and radio
+   buttons
  
  * Added `.serialize()` for encoding forms into URL parameters